+++ src/main/java/com/keepgoing/keepserver/domain/book/domain/repository/BookRepository.java
@@ -1,7 +1,7 @@
-package com.keepgoing.keepserver.domain.book.repository;
+package com.keepgoing.keepserver.domain.book.domain.repository;
 
-import com.keepgoing.keepserver.domain.book.entity.Book;
-import com.keepgoing.keepserver.domain.user.entity.user.User;
+import com.keepgoing.keepserver.domain.book.domain.entity.Book;
+import com.keepgoing.keepserver.domain.user.domain.entity.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
