+++ src/main/java/com/keepgoing/keepserver/domain/book/payload/request/BookDto.java
@@ -1,6 +1,6 @@
 package com.keepgoing.keepserver.domain.book.payload.request;
 
-import com.keepgoing.keepserver.domain.book.entity.enums.BookState;
+import com.keepgoing.keepserver.domain.book.domain.entity.enums.BookState;
 import lombok.Builder;
 
 import java.time.LocalDateTime;
