+++ src/main/java/com/keepgoing/keepserver/domain/book/payload/request/BookRequestDto.java
@@ -1,8 +1,7 @@
 package com.keepgoing.keepserver.domain.book.payload.request;
 
-import com.keepgoing.keepserver.domain.book.entity.enums.BookState;
+import com.keepgoing.keepserver.domain.book.domain.entity.enums.BookState;
 import lombok.Builder;
-import lombok.Getter;
 
 @Builder
 public record BookRequestDto (
